Move around some terrain class (and its test) so we can add a dummy harvestable terrain
